requirements.txt
pytest
selene==2.0.0b5
allure-pytest
python-dotenv==0.20.0
